Replace the laptop example system with a generic Gentoo workstation
This basically just replaces the laptop's minimal system-specific
kernel configuration with the general distribution kernel so the
image can run on any hardware.  (There are additional changes like
supporting multiple userspace video drivers to make it more general
as well.)  As with the other example desktop builds, this supports
an option "nvidia" to include the proprietary video driver, and it
creates the image to run on the same architecture as the build
system.  In this case, though, it tunes binaries for the specific
CPU of the build system.  This can be disabled by removing the two
tuning configuration sections to make a portable image.

Additionally, Arch will now use parallel downloads in its package
manager, and it replaces a flakey mirror that had common timeouts.

The unified UEFI binary has doubled the maximum acceptable kernel
size, up to 32MiB.

The open source Arx Fatalis game engine Arx Libertatis has tagged
its first release in about eight years, so this updates the example
container to use the release archive and stop tracking Git commits.
This allows verifying the code with the maintainer's signing key
for a stronger integrity guarantee.