Replace the 4K display scaling function with a boolean option
The function was intended to configure everything when called in
a system's customize() step, but an option allows writing all of
the different configuration files in more sensible locations.
Maybe the option will be renamed in the future and allow specifying
a numeric scaling factor, but for now there are components with
hard limits on doubling the size (e.g. the Linux console fonts).

This also adds scaling support to LXDM and Xfce.  The Xfce support
is currently relying on being started from LXDM due to it using the
GDK_SCALE environment variable.  (It's not setting the value of
/Gdk/WindowScalingFactor automatically yet.)  Note that other DEs
will inherit LXDM's environment variable as well, so e.g. GNOME
should probably stick to GDM on 4K to avoid scaling twice.