Update example systems for Linux 5.14
The RISC-V example has fixed all kernel and toolchain issues, so it
is using all of the latest versions again.  It also enabled support
for building systemd UEFI applications, but they're not being used
in the final image yet.

Several of the game containers were updated to expect GOG as the
source of the installers, since that's probably where most anti-DRM
people are getting games these days.

The VVVVVV engine build switched to using a release archive instead
of tracking Git commits.  Since the game enables vsync by default
now and that requires libGL, the container supports building with
the "nvidia" option to use the proprietary driver.