Improve the etc git overlay
This adds a post-checkout hook that uses the .gitattributes files
to restore file ownership and modes.  The supported attributes are
owner, group, and mode.  Values are passed directly to the chown,
chgrp, and chmod commands respectively in the order defined, so
be aware of things like Linux chown calls clearing SUID mode bits.
The .gitattributes file is not maintained automatically yet, so it
should be updated when new files are committed that need to save
specific modes/ownership.  For example, saving the first connection
for NetworkManager could add this line:

    /NetworkManager/system-connections/*.nmconnection mode=0600

The post-checkout hook also restores SELinux labels instead of that
happening once during boot.  This should keep things running while
using Git to check out files/branches on a live system.

The SELinux path substitutions from the base installer functions
were moved to the distribution file.  This allows committing system
administrator customizations in the overlay while changes to the
immutable image will still take effect, e.g. if file_contexts.subs
was committed in the overlay and an updated image was built that
enables saving the RPM database in /usr, that new path substitution
update is now visible via the distrubtion file.

Docker will now be socket-activated by default when installed so it
works out of the box.

This fixes Dracut not adding the loop module for 5.14 kernels when
building ramdisk images.

Fedora builds now disable the Cisco repository by default.  It can
be enabled again if needed like any other repository, but it seemed
to just be a slow point for DNF to download more unused metadata.

SELinux labeling was fixed on openSUSE from their new kernel module
settings.  It also fixed EROFS as root when the module blacklist
files were moved.

Gentoo got some minor fixes and removed a bootstrap package build.