Convert an example container from Wine to SCUMMVM
This makes The Longest Journey run in a native SCUMMVM installation
in an Arch container (since it's the only distro that updated its
SCUMMVM package at this point).  The new engine should allow the
game to run on more platforms and with modern display capabilities,
and the container setup should easily translate for packaging any
other SCUMMVM games.

There are also a handful of Gentoo fixes included.  This adds
automatic setup of the powerpc64le architecture, and makes a few
ARM variants default to the new hardened profiles.