Add some built-in root partition slot management
This takes some common pieces of my local scripts and combines them
into something that is hopefully generalized.  It now supports
defining a list of root partition UUIDs instead of specifying a
single target partition.  The list is intended to be all available
root file system slots on the target system's disk, so the build
system can be aware of them for management.  The list of UUIDs can
also be used for setting up repart.d files for initial provisioning
of the entire disk.  A single slot still must be selected for each
build, though.

It has a slightly backwards-incompatible change by making multiple
uses of the -P command-line option add additional partition UUIDs
instead of replacing a single value.  These UUIDs from the CLI are
appended to the new array variable "slots" that system files can
define.  By default, the first slot is selected, so specifying only
a single -P option will have the same effect as it did previously.

The option "slot" can be defined to a capital letter as the index
to select a partition from the "slots" array, where the value "A"
corresponds to the first UUID.  (There isn't really a limit on the
number of slots, but the integer index is just mapped to ASCII
letters, so you probably shouldn't go beyond slot "Z".)  There is a
helper function "get_slot_uuid" that customization scripts can use
to get the UUID from a slot letter.  If not given an argument, it
uses the selected slot by default.

Another helper function "get_next_slot" is used for building on the
target system, to detect the currently booted root slot and return
the letter of the next slot in the list (or the first when running
on the last).  It will return unsuccessfully if the current root
partition is not one of the defined root slots.  For example, a
system configuration file could start with this:

    # Define all root partition UUIDs on the target system.
    slots=(
            6e6c57ac-e9fb-45e1-bf11-77cad8eaddee
            4e1c51f0-560a-4a6f-8eee-b28a32266316
            ...
    )

    # Change build settings based on where this is running.
    options[slot]=$(get_next_slot) && options+=(
            [install_to_disk]=1
            [uefi_path]=/boot/BOOT${options[slot]}.EFI
    ) || options+=(
            [gpt]=1
            [squash]=1
    )

    # Define the usual options, packages, customizations, etc.
    options+=(
            [selinux]=1
            [uefi]=1
            [verity]=1
            ...
    )

When the above is being run on a build system that is not also the
target system, it will create a compressed full-disk image that can
be written directly to disk on the target system to provision it.
When it is being run on the target system (detected by the current
root file system being a partition listed in "slots"), it will
automatically install the resulting root file system image into the
next available partition and place the kernel in the ESP.  This
allows the same script to handle both provisioning and all updates.

There are some other parts that could be included to make this more
"complete" like generating the repart.d files for provisioning, or
adding a method of automatically booting the newest kernel, but the
ways I've handled that seem too system-specific.  I'll push that
stuff when I find a way to make it look more generalized.