Implement IPE in Gentoo
This applies the v7 patch series to the kernel, configures it, and
generates a policy if one isn't provided.  The patch seems to make
the stable kernels fail to set up a root file system, and I haven't
had the time to look into it yet, so it's basically just a build
test at this point.

There are a bunch of fixes to keep rolling release distros running.

Any /.autorelabel files will now be removed after inital labeling.

The openSUSE Factory repo for SELinux packages was dropped since
the regular packages seem to be pretty good now.  It's still using
permissive mode by default, but the only issue I've encountered is
the GDM scripts, which can be worked around with:
    chcon -Rt xsession_exec_t /etc/gdm/*/

The PowerPC Mac mini graphics driver was fixed in new kernels, so
it is built with the latest version again.