Switch PolicyKit to Duktape in Gentoo
This fixes compilation for x32 and any non-Rust platform and fixes
segfaults on PowerPC.  (It also cuts out a lot of build time.)  The
x32 example system is now building the same package set as the rest
of the desktop examples.

RPM verification for the optional repository packages is now fixed
to require valid PGP signatures instead of just digests.

CentOS 9 has fixed ImageMagick, but its networkd is still broken.