Allow changing the SELinux policy
The "selinux" option should now be set to the name of the policy to
install and enable.  If it is still used as a boolean option, it
will use the targeted policy by default for compatibility.

A new boolean option "enforcing" was added to choose whether the
selected policy is being evaluated or enforced.  It is enforced by
default when the distro doesn't have known issues with the targeted
policy (which is currently only Fedora and CentOS).

These options are useful for testing custom policy packages, but
they could also be used with the other distro policies.  For
example, use "-o enforcing= -o selinux=mls" to test the MLS policy
in permissive mode.