Add Ubuntu 22.04
A new helper function was added to enable a PPA in Ubuntu, and the
desktop example system uses it to install a real Firefox package
now.  This function takes the name of the PPA as an argument and
expects its signing GPG key to be passed on stdin.  The desktop
example limits the PPA to prioritize only Firefox above the other
Ubuntu packages, so nothing else is affected.

All Ubuntu example builds are updated to 22.04.

All example systems have been switched to wireplumber by default.