Automatically enroll Secure Boot certificates for QEMU/OVMF
A new option uefi_vars is supported to specify a colon-separated
list of four paths on the host to OVMF files: the minimal variable
template, the BIOS file, a UEFI shell, and the EnrollDefaultKeys
program.  All paths are optional, and any missing files will be
used from the target distro (if available).  Setting uefi_vars as a
boolean option will use all files from the target distro.

Be aware that there is no guarantee that a variable template will
work with BIOS from a different OVMF build.  This option should
always be given at least the first two paths to specify the OVMF
pflash files that will be used on the target hypervisor system.

The initial output directory creation was moved to a new function
create_working_directory.  This function could be overridden for
more control of where the build runs.

CentOS 8 has been resurrected so that it can build with the final
state of its packages in 8.5, although it will not receive updates
due to being dead upstream.