Update some things for Linux 5.18
A few example Gentoo systems were getting out of date due to kernel
breakage, so this fixes everything.  The chromebook's broken panel
driver was fixed with 5.18.  The yoga wireless driver is patched to
build again.  Gentoo SELinux is using a patched audit version so
kernel headers don't have to be downgraded.

The Fedora example wireless driver branch is switched to support
building with 5.18.

Gentoo's new global kernel module compression behavior should work
with zstd everywhere now.

The IPE patch is updated to v8 in Gentoo, but it needs further
adjustments to apply to stable kernels.

The unified UEFI binary will now automatically include a file named
devicetree.dtb if it is present, so a device tree can be chosen for
UEFI booting on systems with atypical architectures.

A new example game container for X-COM: UFO Defense has been added
to demonstrate containerizing software running in DOSBox.