Fix Gentoo and update to Linux 6.1
All of the Gentoo examples were fixed and tested to work on real
hardware with Linux 6.1.

The profile for x86_64 was switched to the 23.0 version.  It also
exists for ARM, but the profile is broken for all non-x86 platforms
since it uses an x86-specific flag that outputs warnings elsewhere
which break configure tests.  It had to be reverted to 17.1 for the
buildroot since its UsrMerge switch breaks stage3s.  It conflicts
with Firefox's ELF hack, too.

The bootstrapping packages were reworked, and now systemd is built
separately to satisfy circular dependencies on a udev library.  New
circular dependencies should be solved more easily with this.