Fix some more package sets in Gentoo
This allows installing stuff related to the Enlightenment and Sway
desktops, specifically for Wayland.  The handful of Gentoo example
systems that support DRM now install Sway to test things in a pure
Wayland environment.  (Important packages like Firefox and VLC are
working.)  Those examples will probably be switched to Wayland by
default as soon as Xfce's support is stable.  Enlightenment fixes
might be dropped since the packages are too huge to justify
installing to test applications.

Qt 6 is also fixed so it can be cross-compiled anywhere now.  The
Qt 5 package fixes have been isolated so they don't conflict with
newer versions.

The default x86 profile was switched to 23.0.  The 23.0 profile is
still broken on all non-x86 architectures as mentioned in the last
commit, so it's being used everywhere it can be for now.