Add Ubuntu 23.10
Piles of fixes have accumulated in this.  There are still several
broken things, notably Linux 6.5 broke using PARTUUID with dm-init.

UKI generation is now handled as in dracut, where it calculates the
next consecutive aligned address to add sections so that it's more
robust for handling layout changes from systemd.

The VTMB example container had some new Wine issues with the
previous Ubuntu upgrade, so it was replaced with another example
container to build the original version of The Sims.  It's probably
more helpful as an example since it uses different installers.