Add Fedora 41 and Ubuntu 24.10
All builds will now save a snapshot of their code and options in a
new script "rebuild.sh" in the output directory.  This was intended
to help debugging old builds, but it can also be run as a single
self-contained script to build a system without this repository.

Fedora builds have switched to the minimal container images since
microdnf is the same as dnf5 now.  Since the minimal image requires
installing more packages in the buildroot, this seemed like a fine
time to switch the default behavior to keep the packages cached so
they don't need to be redownloaded for the main install.

CentOS 9 was also switched to its minimal container version since
that doesn't use microdnf, and it's also caching packages.  CentOS
7 builds now point at the archive server since it's EOL.

The Fedora example system is no longer resisting Wayland.  It seems
to work decently enough lately to make the switch.  All of the
example container builds were updated to support running on a host
that uses Wayland, however they'll require Xwayland since the stuff
in the containers is still all X11 clients.

I haven't fixed all the Gentoo breakage in this commit, so don't
use it.  Its IPE patching was removed since it was merged for Linux
6.12.  Soon we'll finally be able to enforce that programs must be
read from a trusted verity device.  It's a start.